requirements.txt
websockets==12.0
speechbrain==1.0.0
pyannote.audio==3.3.2
asyncio==3.4.3
sentence-transformers==2.7.0
transformers==4.40.2
faster-whisper==1.0.2
torchvision~=0.18.0
torch~=2.3.0
python-dotenv==1.0.1
structlog==25.2.0
asgi-correlation-id==4.3.4
